+++ admin/config-overrides.js
@@ -1,3 +1,4 @@
+const webpack = require( 'webpack' );
 const { compose } = require( 'react-app-rewired' );
 const rewireStyledComponents = require( 'react-app-rewire-styled-components' );
 const rewireSVG = require( 'react-app-rewire-svg-react-loader' );
@@ -12,17 +13,21 @@ module.exports = function ( config, env ) {
 
 	if ( env === 'production' ) {
 		config.plugins.push( new DynamicPublicPathPlugin( {
-			externalGlobal: 'window.HM.Workflows.BuildURL', //Your global variable name.
+			externalGlobal: 'window.HM.Workflows.BuildURL',
 			chunkName: 'hm-workflows',
 		} ) );
 	}
 
+	// Override entry to customise entry file name.
 	config.entry = { 'hm-workflows': config.entry };
 
+	// Avoid code splitting conflicts.
 	config.output.jsonpFunction = 'hmWorkflowsJSONP';
 
-	// Isolate styled components.
-	process.env.SC_ATTR = 'hm-workflows';
+	// Set SC_ATTR env var.
+	config.plugins.push( new webpack.EnvironmentPlugin( {
+		SC_ATTR: 'data-styled-components-hm-workflows',
+	} ) );
 
 	return rewires( config, env );
 }